+++ basyx.aasrepository/basyx.aasrepository-feature-registry-integration/src/main/java/org/eclipse/digitaltwin/basyx/aasrepository/feature/registry/integration/RegistryIntegrationAasRepositoryConfiguration.java
@@ -27,6 +27,7 @@
 package org.eclipse.digitaltwin.basyx.aasrepository.feature.registry.integration;
 
 import org.eclipse.digitaltwin.basyx.aasregistry.client.api.RegistryAndDiscoveryInterfaceApi;
+import org.eclipse.digitaltwin.basyx.aasrepository.AasRepository;
 import org.eclipse.paho.client.mqttv3.MqttException;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
@@ -34,13 +35,18 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
+/**
+ * Configuration for {@link AasRepository} with AasRegistry
+ * 
+ * @author danish
+ */
 @Configuration
 @ConditionalOnExpression("!T(org.springframework.util.StringUtils).isEmpty('${basyx.aasrepository.feature.registryintegration:}') && !T(org.springframework.util.StringUtils).isEmpty('${basyx.aasrepository.externalurl:}')")
 public class RegistryIntegrationAasRepositoryConfiguration {
 
 	@Bean
 	@ConditionalOnMissingBean
-	public AasRepositoryRegistryLink mqttClient(@Value("${basyx.aasrepository.feature.registryintegration}") String registryBasePath, @Value("${basyx.aasrepository.externalurl}") String aasRepositoryURL) throws MqttException {
+	public AasRepositoryRegistryLink getAasRepositoryRegistryLink(@Value("${basyx.aasrepository.feature.registryintegration}") String registryBasePath, @Value("${basyx.aasrepository.externalurl}") String aasRepositoryURL) throws MqttException {
 	
 		return new AasRepositoryRegistryLink(new RegistryAndDiscoveryInterfaceApi(registryBasePath), aasRepositoryURL);
 	}
